Added delete account functionality

+++ public/javascript/settings.js
@@ -74,25 +74,33 @@ function resetPassword() {
 
 
 function deleteAccount() {
-    console.log("Delete account :)");
+    console.log("Delete account :");
     firebase.auth().onAuthStateChanged(function (user) {
         if (user) {
             var userId = user.uid;
-            var canBeDeleted = false;
             firebase.database().ref('users/' + userId + '/applications').once('value').then(function (snapshot) {
                 if (snapshot.val() == null) {
-                    canBeDeleted = true;
+                    console.log("inside if");
+                    firebase.auth().currentUser.delete().then(function () {
+                        firebase.database().ref("users/" + userId).remove();
+                        alert('Your account has been successfully deleted!');
+                        window.location.href = "login";
+                    });
                 } else {
+                    console.log("inside else");
                     snapshot.forEach(function (applicationId) {
                         var department = applicationId.child("department").val();
                         var term = applicationId.child("term").val();
                         firebase.database().ref("applications/" + term + "/" + department + "/" + applicationId).once('value').then(function (snapshot) {
-                            canBeDeleted = !(snapshot.child("gradschoolControl/isVerified").val());
-                            firebase.auth().currentUser.delete().then(function () {
-                                firebase.database().ref("users/" + userId).remove();
-                                alert('Your account has been successfully deleted!');
-                                window.location.href = "login";
-                            });
+
+                            if (snapshot.child("gradschoolControl/isVerified").val() === false) {
+                                firebase.auth().currentUser.delete().then(function () {
+                                    firebase.database().ref("users/" + userId).remove();
+                                    alert('Your account has been successfully deleted!');
+                                    window.location.href = "login";
+                                });
+                            }
+
                         }).catch(function (error) {});
                     });
                 }
